Separate out queue
